Deprecate `zntrack.zn.<field>` and `zntrack.dvc.<field>` (#722)
* deprecated zntrack.zn and zntrack.dvc

* deprecate 'zntrack.zn' and 'zntrack.dvc' modules

* bump zntrack version to 0.7.0

* update imports
