+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "ZnTrack"
-version = "0.6.3"
+version = "0.7.0"
 description = "Create, Run and Benchmark DVC Pipelines in Python"
 authors = ["zincwarecode <zincwarecode@gmail.com>"]
 license = "Apache-2.0"
@@ -20,6 +20,7 @@ dot4dict = "^0.1.1"
 zninit = "^0.1.11"
 znjson = "^0.2.2"
 znflow = "^0.1.14"
+typing-extensions = "^4.8.0"
 
 
 [tool.poetry.urls]
