+++ zntrack/fields/fields.py
@@ -1,6 +1,7 @@
 """Fields that are used to define Nodes."""
-from zntrack.fields import dvc, zn
 from zntrack.fields.dependency import Dependency
+from zntrack.fields.dvc.options import DVCOption, PlotsOption
+from zntrack.fields.zn.options import Output, Params, Plots
 
 # Serialized Fields
 
@@ -16,7 +17,7 @@ def outs():
             and stored in the node working directory.
             see https://dvc.org/doc/command-reference/stage/add#-o
     """
-    return zn.outs()
+    return Output(dvc_option="outs", use_repr=False)
 
 
 def metrics():
@@ -31,22 +32,24 @@ def metrics():
             see https://dvc.org/doc/command-reference/stage/add#-M
 
     """
-    return zn.metrics()
+    return Output(dvc_option="metrics-no-cache")
 
 
-def params(*data):
+def params(*args, **kwargs):
     """Define a Node Parameter.
 
     Parameters
     ----------
-    data: any
+    args: any
         A data object that is used as a parameter.
         Typically, this should be a string or number.
         The object is serialized and deserialized by ZnTrack
         and stored in params.yaml.
         see https://dvc.org/doc/command-reference/stage/add#-p
+    kwargs: dict
+        Additional keyword arguments.
     """
-    return zn.params(*data)
+    return Params(*args, **kwargs)
 
 
 def deps(*data):
@@ -63,12 +66,12 @@ def deps(*data):
     return Dependency(*data)
 
 
-def plots(*data, **kwargs):
+def plots(*args, **kwargs):
     """Define a Node Plot.
 
     Parameters
     ----------
-    data: pd.DataFrame
+    args: pd.DataFrame
         A pandas DataFrame that is used as a plot.
         The object is serialized and deserialized by ZnTrack
         and stored in the node working directory.
@@ -77,73 +80,87 @@ def plots(*data, **kwargs):
         Additional keyword arguments that are used for plotting.
 
     """
-    return zn.plots(*data, **kwargs)
+    return Plots(*args, **kwargs)
 
 
 # Path Fields
 
 
-def outs_path(*path):
+def outs_path(*args, dvc_option="outs", **kwargs):
     """Define a Node Output.
 
     Parameters
     ----------
-    path: str|Path
+    args: str|Path
         A file or directory that is generated by the Node.
         see https://dvc.org/doc/command-reference/stage/add#-o
+    dvc_option: str, default="outs"
+        The DVC option to use for this field.
+    kwargs: dict
+        Additional keyword arguments.
 
     """
-    return dvc.outs(*path)
+    return DVCOption(*args, dvc_option=dvc_option, **kwargs)
 
 
-def metrics_path(*path):
+def metrics_path(*args, dvc_option="metrics", **kwargs):
     """Define a Node Metric.
 
     Parameters
     ----------
-    path : str|Path
+    args : str|Path
         A file that is used by DVC as a metric, such as *.json
         see https://dvc.org/doc/command-reference/stage/add#-M
+    dvc_option: str, default="metrics"
+        The DVC option to use for this field.
+    kwargs: dict
+        Additional keyword arguments.
 
     """
-    return dvc.metrics(*path)
+    return DVCOption(*args, dvc_option=dvc_option, **kwargs)
 
 
-def params_path(*path):
+def params_path(*args, **kwargs):
     """Define a Node Parameter.
 
     Parameters
     ----------
-    path : str|Path
+    args : str|Path
         A file that is used by DVC for reading parameters.
         This includes typically json or yaml files.
         see https://dvc.org/doc/command-reference/stage/add#-p
+    kwargs: dict
+        Additional keyword arguments.
     """
-    return dvc.params(*path)
+    return DVCOption(*args, dvc_option="params", **kwargs)
 
 
-def deps_path(*path):
+def deps_path(*args, **kwargs):
     """Define a Node Dependency.
 
     Parameters
     ----------
-    path : str|Path
+    args : str|Path
         A file or directory that is defined as a dependency to the Node.
         see https://dvc.org/doc/command-reference/stage/add#-d
+    kwargs: dict
+        Additional keyword arguments.
 
     """
-    return dvc.deps(*path)
+    return DVCOption(*args, dvc_option="deps", **kwargs)
 
 
-def plots_path(*path, **kwargs):
+def plots_path(*args, dvc_option="plots", **kwargs):
     """Define a Node Plot.
 
     Parameters
     ----------
-    path : str|Path
+    args : str|Path
         A file or directory that is defined as a plot to the Node.
         see https://dvc.org/doc/command-reference/stage/add#--plots
+    dvc_option: str, default="plots"
+        The DVC option to use for this field.
     kwargs: dict
         Additional keyword arguments that are used for plotting.
     """
-    return dvc.plots(*path, **kwargs)
+    return PlotsOption(*args, dvc_option=dvc_option, **kwargs)
